feat: Restructure the program
This commit erases the '-v' flag and the verbosity mode and modify the
'-x'  interception. Now it has to be the last option.

The subcommand interpretation are also modified to allow portability to
repositories paths. Now runs, e.g.: "... -x 'cd .git && pwd'", because
the 'cd' is interpreted locally in the repositories located subshell.